Add bits files
This commit adds the *.bits files to the source (in the bits directory),
and adds the tool and Makefile gunk necessary to construct human-readable
presentations from those files.  This also narrows down the targets to
pass on to Sphinx (otherwise make tries to invoke sphinx for "bits") and
replaces the "help" target with an "all" target that goes to "html".