Add the beginnings of the protocol basics
This commit contains a discussion of the basics of the Humboldt protocol,
describing the carrier protocol, extensions, round trip time calculation,
and a discussion of retransmissions and acknowledgments.  This commit also
introduces a bibliography and a list of configuration variables, with the
two variables used by the retransmission logic.  Finally, the extension
protocol format is augmented with a new HOP bit, indicating a hop-by-hop
extension that should be dropped before forwarding the frame.