Tweak table of protocols
This commit tweaks the table of protocols: rename the "Number" column to
"Protocol Number" for clarity; remove the "Uses xxx" columns; add "Sent
From" and "Sent To" columns.  It also adopts the same convention used to
describe variables by adding a one-row table at the beginning of the
Protocol Versioning and Negotiation section of basics.rst.