Describe the ping protocol
This commit adds a discussion of the ping protocol to the support
protocols chapter.  It also enhances the protobuf files with more
descriptive comments to improve understanding outside the context of the
text.  Finally, the extensions.proto is linked from the basics document's
description of the protobuf encoding, and the message names are added to
the node ID protocol section.