Discuss debouncing
This commit moves the discussion of debouncing out of both of the
sections where it occurs and into its own subsection in the basics
document.  It also describes the configuration variables that are used
when applying debouncing to link-state topics.  Finally, a minor typo
with respect to the LS_HORIZON configuration is corrected.